fix metrics and llm judge

+++ src/llm_utils.py
@@ -86,4 +86,4 @@ x/5
             _log_llm(messages, error=str(e))
             raise e
             
-    return int(resp_content.split("### Total Score:")[1].strip().split("/")[0])
+    return float(resp_content.split("### Total Score:")[1].strip().split("/")[0])
